Implemented code in heaven art

+++ Assets/Scripts/Level4/CloudsLeanTween.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class CloudsLeanTween : MonoBehaviour
 {
     [SerializeField] LeanTweenType type;
-    [SerializeField] int _duration;
+    [SerializeField] float _duration;
     [SerializeField] float _moveTo;
     // Start is called before the first frame update
     void Start()
